requirements.txt
numpy>=1.19.5
scikit-learn>=0.24.2
matplotlib>=3.4.3
seaborn>=0.11.2
tqdm>=4.62.3
pyyaml>=5.4.1